% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-eval.R
\name{cluster_eval}
\alias{cluster_eval}
\alias{cluster_eval_}
\title{Evaluate an expression on a cluster}
\usage{
cluster_eval(cluster, expr)

cluster_eval_(cluster, expr)
}
\description{
Evaluate an expression on a cluster
}
